Gate install of tmpfiles configuration
The vhost-user tmpfiles configuration is only applicable in deployments
using libvirt/kvm with nova-compute.

Ensure appropriate user and group exists before installing tmpfiles.d
configuration.

Closes-Bug: 1792414

+++ hooks/neutron_ovs_utils.py
@@ -67,6 +67,8 @@ from charmhelpers.core.host import (
     service_running,
     CompareHostReleases,
     init_is_systemd,
+    group_exists,
+    user_exists,
 )
 
 from charmhelpers.fetch import (
@@ -445,7 +447,10 @@ def enable_ovs_dpdk():
 
 def install_tmpfilesd():
     '''Install systemd-tmpfiles configuration for ovs vhost-user sockets'''
-    if init_is_systemd():
+    # NOTE(jamespage): Only do this if libvirt is actually installed
+    if (init_is_systemd() and
+            user_exists('libvirt-qemu') and
+            group_exists('kvm')):
         shutil.copy('files/nova-ovs-vhost-user.conf',
                     '/etc/tmpfiles.d')
         subprocess.check_call(['systemd-tmpfiles', '--create'])
